[fix] Ensure transform dir is present for install
Generated apps’ post-install.xql files contain a local:generate-code() function that generates a fresh copy of all transform scripts during installation, so the files can be excluded source control (i.e., by adding the transform directory to the .gitignore file). But the transform directory must be present when the files are created. Thus, to be safe, this creates the required “transform” collection. If the collection is already present, it’s not affected.